Moved avatar-list to better location
