+++ src/app/app.module.ts
@@ -17,7 +17,7 @@ import { UserRegisterComponent } from './users/register/user-register.component'
 import { ReactiveFormsModule } from '@angular/forms';
 import { UserProfileComponent } from './users/profile/user-profile.component';
 import { ProjectDetailsComponent } from './projects/project-details/project-details.component';
-import { UserAvatarListComponent } from './user-avatar-list/user-avatar-list.component';
+import { UserAvatarListComponent } from './users/user-avatar-list/user-avatar-list.component';
 
 @NgModule({
   declarations: [
