+++ src/app/users/user-avatar-list/user-avatar-list.component.scss
